+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## master
+
+Changes:
+
+- Ensure `BTreeSet` has the correct decoded length
+
+
 ## 7.9.1 Feb 20, 2022
 
 Contributed: